SECOND COMMIT - Refactored, working, but glitched
The code here has been refactored a bit to be better. In particular, the
question-asking functions have been made more general and effective,
though they are still somewhat limited in their helper methods to only
two options. However, they have been replaced with enums for better usages
and the utilities have been shaped around that new infrastructure. The
questions work as they did previously, and even now catch when there are
non-int inputs into the keyboard. The test game still works.

However, there are still glitches, namely in summing up the player and
dealer hands. This could be a result of the checking ace functions, which
were added to allow the player to choose ace as 1 or 11 and the dealer
to choose whichever does not result in bust. These will need to be revised
and this will likely be helped with a better card system where their types
are enums rather than indexed integers or values.

ISSUES RESOLVED:
1. Player can now choose ace as 11 or 1
2. Question asking methods have been made more general
3. No more errors from choice function

ISSUES REMAINING:
1. Glitches/logical errors in summing of cards
2. Players can still only hold 3 cards total
3. Code can be made clearer/less verbose
4. Still need way to continue after one hand
5. Can use loops instead of recursion
6. Card logic needs rework - enums, better summing, better data storing
7. Others I can't think of but will make issues for.

Thank you!