card equivalence
modify card and deck slightly, so that
cards now override equality to be
based on suit and rank. additionally, updated deck to only add cards
during initializing if the deck does not have
those cards in it already. also made
a draw card function which is not currently
used but can be used to replace some of the
functionality in tester.