piles and hands
modifying the class inheritance a bit so that now we
have piles and hands. both decks and hands are a type
of CardPile. general functions have been added in pile,
with specific functions in deck and hand. also modified
the code slightly in tester to leverage some of these
changes. there is more to be done to fully take advantage
of the new changes. for example, we no longer need
the choice for ace.