rough round logic
this is a work in progress at the moment so not committing
it to my primary branch. the goal will be to get all of the
round logic working so that i can merge it into my primary
branch before merging it into main. committing here to
save my current flow and help me remember the direction
i am trying to move things towards. the code is not
yet working.