round logic
adding logic to play the round and determine the winner.
a round will continue until all of the players are standing
or busted -- or if a player reaches blackjack.

also, added busted as a choice enum to handle that case which
is a bit unique to standing or hitting. also modified the
logic for player choices so that it is now a function on
player, that gets overridden by its subclasses. may need to
add the override decorator, will have to check.

so far, this code has not really been tested. but after a few
more changes i should be able to modify the tester code to
test the logic here. once it works, will merge into my branch.