winners
adding logic in the round for determining who the winners
of the round are. there could hypothetically be multiple
winners in the case of a tie, but the dealer's win will
always take precedence. the win stats and win printing
is handled by the player class itself now. i think the
next step is pretty much going to be updating the tester
logic to incorporate the new changes and test what has
been committed, which is currently untested -- so there
are likely a few bugs.

then, we can look at adding comments and tests, etc.