comments and refactoring
refactored the code structure so that now everything
is divided up into packages that are a bit more
specific to their purpose, so that the code files
are a bit more cleanly structured and easier to
maintain and reference. however, i am not a java
developer so i am not sure if this structure makes
sense. i had to make some functions public that
were previously protected as a result (an outside
package cannot access protected functions), so
we may want to rethink this as we consider how
we have a better strategy for protecting the
functions, so on.

additionally, added some comments including some
todos. some general todos:
- clean up the code a bit, including reordering
  functions to have some readability and consistency
- add more comments
- figure out the proper scope for functions, such as
  whether they should be protected, public, private, etc
- address comment todos
- implement logic for more strategies such as split, etc
- generalize and abstract logic where it can be (such
  as when prompting for user input on continuing or hitting)
- reorganize code a bit so that things like choice and
  gamechoice arent named weirdly similarly, or make them
  subclasses of something else -- figure out a good approach
  here
- reorganize project structure so packages make more sense
  and are java-ic
- fix any bugs that arise
- make print statements cleaner and more clear
- allow players to input their names
- access a player's hand operations through the player,
  instead of getting their hand to operate on it (?)
- allow for more complex NpcPlayers that have better
  decision logic
- allow for larger "blackjack" decks
- make player state an enum
- add error handling
- add unit tests
and beyond!