refactoring (#8)
* card equivalence

modify card and deck slightly, so that
cards now override equality to be
based on suit and rank. additionally, updated deck to only add cards
during initializing if the deck does not have
those cards in it already. also made
a draw card function which is not currently
used but can be used to replace some of the
functionality in tester.

* piles and hands

modifying the class inheritance a bit so that now we
have piles and hands. both decks and hands are a type
of CardPile. general functions have been added in pile,
with specific functions in deck and hand. also modified
the code slightly in tester to leverage some of these
changes. there is more to be done to fully take advantage
of the new changes. for example, we no longer need
the choice for ace.

* rough round logic (#9)

* rough round logic

this is a work in progress at the moment so not committing
it to my primary branch. the goal will be to get all of the
round logic working so that i can merge it into my primary
branch before merging it into main. committing here to
save my current flow and help me remember the direction
i am trying to move things towards. the code is not
yet working.

* player tostring

bringing dealer and player makechoice up to parity
with each other so they will be easier to call
during a round. additionally, added some tostrings
so that players can see the visible cards of
other players more easily. in the future, will do
things such as allowing for npc players other
than the dealer, and will do some more logic for
these pieces.

* round logic

adding logic to play the round and determine the winner.
a round will continue until all of the players are standing
or busted -- or if a player reaches blackjack.

also, added busted as a choice enum to handle that case which
is a bit unique to standing or hitting. also modified the
logic for player choices so that it is now a function on
player, that gets overridden by its subclasses. may need to
add the override decorator, will have to check.

so far, this code has not really been tested. but after a few
more changes i should be able to modify the tester code to
test the logic here. once it works, will merge into my branch.

* winners

adding logic in the round for determining who the winners
of the round are. there could hypothetically be multiple
winners in the case of a tie, but the dealer's win will
always take precedence. the win stats and win printing
is handled by the player class itself now. i think the
next step is pretty much going to be updating the tester
logic to incorporate the new changes and test what has
been committed, which is currently untested -- so there
are likely a few bugs.

then, we can look at adding comments and tests, etc.

* the game runs

now the game runs using the round and game logic that has
been defined, removed some old code that was causing tester
to fail. had to make a lot of updates to get my code working.

there are still some issues, for example:
- the dealer will not continue hitting once they get to 17
- need to add better printing around players busting
- need to do some general cleanup
etc

* comments and refactoring

refactored the code structure so that now everything
is divided up into packages that are a bit more
specific to their purpose, so that the code files
are a bit more cleanly structured and easier to
maintain and reference. however, i am not a java
developer so i am not sure if this structure makes
sense. i had to make some functions public that
were previously protected as a result (an outside
package cannot access protected functions), so
we may want to rethink this as we consider how
we have a better strategy for protecting the
functions, so on.

additionally, added some comments including some
todos. some general todos:
- clean up the code a bit, including reordering
  functions to have some readability and consistency
- add more comments
- figure out the proper scope for functions, such as
  whether they should be protected, public, private, etc
- address comment todos
- implement logic for more strategies such as split, etc
- generalize and abstract logic where it can be (such
  as when prompting for user input on continuing or hitting)
- reorganize code a bit so that things like choice and
  gamechoice arent named weirdly similarly, or make them
  subclasses of something else -- figure out a good approach
  here
- reorganize project structure so packages make more sense
  and are java-ic
- fix any bugs that arise
- make print statements cleaner and more clear
- allow players to input their names
- access a player's hand operations through the player,
  instead of getting their hand to operate on it (?)
- allow for more complex NpcPlayers that have better
  decision logic
- allow for larger "blackjack" decks
- make player state an enum
- add error handling
- add unit tests
and beyond!

* more todos

added a couple more todos that i think could make
things even better moving forward.