Determine Winner Changes for Functionality
The determine winner function is changed yet again and this seems to
work? It instead looks at the dealer and sees if the dealer should be
in either of the winner arraylists. If so then logic is done, to make
him win. If not, then logic is done to determine player winners.

This should work with multiple players (tester is changed to have 2
human players), and this should work for edge cases, but idk for sure.