Small commit - last minute changes
Sorry for the small commit, just realized in determineWinners we don't
need to loop through the winners arraylist and check for the dealer
because the dealer will never be in there with players; if he's in
there, he wins automatically.

Then just added comments to clarify return statements etc. Hopefully all
still works. Last change I'll make before the review.