DetermineWinners Refactored and Fixed
DetermineWinners() method has been updated and refactored to be much
cleaner and smaller as well as hopefully fully functional with any
number of players. First it initializes the winners arraylist with a 0
for maxScore, then loops through all players. It prints their info,
checks if they are busted, and if not, checks if their handValue is
higher than the player that came previously. If so, they are the only
winner. If equal, they get added with previous player(s).

Now we allow for ties with the dealer even if the dealer has blackjack
for future betting/push possibilities.

Tester has also been changed to test for a game with 3 players, which
has helped with testing the different cases.