Round.java Bug Fixes & Refactoring (#12)
* Players Validation Error Handling

Added validatePlayers method in Round for error handling if dealer ever
 ends up in theArrayList of players. Added new Exception to extend
 RuntimeException that throws PlayerValidationException, which closes
 the program with a custom error message if the dealer ever ends up in
  the players Array.

* New Round Constructor

Added a new Round constructor for one dealer and one player which
creates players arraylist and just adds the one player passed in as a
parameter, for more flexible round construction.

* Determine Winners Refactoring and Print Statement Additions

Refactored the determineWinners method to check if blackjack regardless
of dealer status, then check higher hand regardless of dealer status,
then afterward iterate through the winners list. If dealer is present
in that list, dealer wins and player no longer is in the winners list.
Then returns the list for proper functionality with other methods.
Functionality went from ~27 lines without comments to ~22 with comments.

Also added print lines for printing out the winners after the "here
they are:" which originally had nothing after them. Iterate through the
winners and print their names, if there are winners; otherwise print
"None."

Added line comments for things that have been changed.

* Determine Winner Changes for Functionality

The determine winner function is changed yet again and this seems to
work? It instead looks at the dealer and sees if the dealer should be
in either of the winner arraylists. If so then logic is done, to make
him win. If not, then logic is done to determine player winners.

This should work with multiple players (tester is changed to have 2
human players), and this should work for edge cases, but idk for sure.

* Small commit - last minute changes

Sorry for the small commit, just realized in determineWinners we don't
need to loop through the winners arraylist and check for the dealer
because the dealer will never be in there with players; if he's in
there, he wins automatically.

Then just added comments to clarify return statements etc. Hopefully all
still works. Last change I'll make before the review.

* DetermineWinners Refactored and Fixed

DetermineWinners() method has been updated and refactored to be much
cleaner and smaller as well as hopefully fully functional with any
number of players. First it initializes the winners arraylist with a 0
for maxScore, then loops through all players. It prints their info,
checks if they are busted, and if not, checks if their handValue is
higher than the player that came previously. If so, they are the only
winner. If equal, they get added with previous player(s).

Now we allow for ties with the dealer even if the dealer has blackjack
for future betting/push possibilities.

Tester has also been changed to test for a game with 3 players, which
has helped with testing the different cases.