Dealer Hit Logic + Minor code fixes
Modified dealer hit logic in the Game.java file to align with actual
Blackjack and prevent dealer from busting so often; dealer hits when
hand value is below 16, but stands when 17 or higher.

Also removed a space in Round.java.

Also refactored enum names to be all capitals to fit with Java naming
convention, which modified many files.