NPCPlayer Added with Basic AI and PlayerState Enum Added w/o function
NPCPlayer class was added using the UUID String name constructor from
the player class, though that can be changed if the names are too clunky
and awkward. In the NPCPlayer function they have their own basic AI to
act the exact same as the dealer. Perhaps in future iterations, we can
add different "personalities" to AIs to have them act differently,
particularly when betting is involved. As well as difficulties.

PlayerState enum was also added to track player state rather than using
booleans. However, they have not been implemented in the actual code yet
and are just added to the initialization of the instance and the
startGame() method.

Future commits will be adding functionality to both NPCs and the enum.